수정

+++ studyroom/cron.py
@@ -3,5 +3,6 @@ import datetime
 
 def schedule_monday():
     real_time = datetime.datetime.now().weekday()
-    q = Reserve.objects.all()
-    q.delete()
+    if real_time == 0:
+        q = Reserve.objects.all()
+        q.delete()
